\name{exrates}
\alias{exrates}
\docType{data}
\title{ wright's Exchange Rates Data}
\description{
  The data set used in Wright (2001) as an application, weekly from August, 7, 1974 to May 29 1996
}
\usage{data(exrates)}
\format{
  A data frame with 1139 observations on the following 5 variables.
  \describe{
    \item{\code{ca}}{a numeric vector, Canadian Dollar}
    \item{\code{dm}}{a numeric vector, Deutch Mark}
    \item{\code{ff}}{a numeric vector, French Franc}
    \item{\code{uk}}{a numeric vector, UK Pound}
    \item{\code{jp}}{a numeric vector, Japanese Yen}
  }
}

\references{
  WRIGHT,J.H.,2000,Alternative Variance-Ratio Tests Using Ranks and Signs, Journal of Business & Economic Statistics, 18, 1-9.
}
\examples{
data(exrates)
}
\keyword{datasets}
