\name{vrtest-package}
\alias{vrtest-package}
\alias{vrtest}
\docType{package}
\title{
Variance Ratio tests and spectral shape tests for Weak-form Market Efficiency
}
\description{
A collection of variance ratio and spectral shapte tests
}
\details{
\tabular{ll}{
Package: \tab vrtest\cr
Type: \tab Package\cr
Version: \tab 0.94\cr
Date: \tab 2009-06-17\cr
License: \tab GPL version 2 or newer\cr
}
}
\author{
Jae H. Kim

Maintainer: Jae H. Kim <J.Kim@latrobe.edu.au>

}
\keyword{  htest  }