\name{vrtest-package}
\alias{vrtest-package}
\alias{vrtest}
\docType{package}
\title{
Variance Ratio tests and other tests for Martingale Difference Hypothesis
}
\description{
A collection of variance ratio and spectral shapte tests
}
\details{
\tabular{ll}{
Package: \tab vrtest\cr
Type: \tab Package\cr
Version: \tab 0.96\cr
Date: \tab 2014-04-22\cr
License: \tab GPL-2 \cr
}
}
\author{
Jae H. Kim

Maintainer: Jae H. Kim <J.Kim@latrobe.edu.au>

}
\keyword{  htest  }