\name{vrtest-package}
\alias{vrtest-package}
\alias{vrtest}
\docType{package}
\title{
Variance Ratio Tests for Weak-form Market Efficiency
}
\description{
A collection of the R functions for variance ratio tests
}
\details{
\tabular{ll}{
Package: \tab vrtest\cr
Type: \tab Package\cr
Version: \tab 0.8\cr
Date: \tab 2008-02-21\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Jae H. Kim

Maintainer: Jae H. Kim <Jae.Kim@buseco.monash.edu.au>

}
\keyword{  htest  }