\name{vrtest-package}
\alias{vrtest-package}
\alias{vrtest}
\docType{package}
\title{
Variance Ratio Tests for Weak-form Market Efficiency
}
\description{
A collection of the R functions for variance ratio tests
}
\details{
\tabular{ll}{
Package: \tab vrtest\cr
Type: \tab Package\cr
Version: \tab 0.7\cr
Date: \tab 2007-06-12\cr
License: \tab GPL version 2 or newer\cr
}

}
\author{
Jae H. Kim

Maintainer: Jae H. Kim <Jae.Kim@buseco.monash.edu.au>

}
\keyword{  htest  }