\name{vrtest-package}
\alias{vrtest-package}
\alias{vrtest}
\docType{package}
\title{
Variance Ratio tests and other tests for Martigale Difference Hypothesis
}
\description{
A collection of variance ratio and spectral shapte tests
}
\details{
\tabular{ll}{
Package: \tab vrtest\cr
Type: \tab Package\cr
Version: \tab 0.95\cr
Date: \tab 2010-11-01\cr
License: \tab GPL version 2 or newer\cr
}
}
\author{
Jae H. Kim

Maintainer: Jae H. Kim <J.Kim@latrobe.edu.au>

}
\keyword{  htest  }